// src/types/IpData.ts
export interface IPData {
  ip: string
  city?: string
  region?: string
  country?: string
  country_code?: string
  timezone?: string
  isp?: string
  org?: string
  as?: string
  latitude?: number
  longitude?: number
  postal?: string
  currency?: string
  languages?: string
  continent_code?: string
  utc_offset?: string
  calling_code?: string
  version?: string
  network?: string
}